# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mindful Productivity API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task operations
  - name: moods
    description: Mood tracking operations
  - name: sleep
    description: Sleep Support operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /tasks/summary:
    get:
      operationId: getTaskSummary
      tags: [tasks]
      summary: Get task summary counts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Task summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSummary"
  /moods:
    get:
      operationId: listMoods
      tags: [moods]
      summary: List mood entries
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of mood entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoodEntry"
    post:
      operationId: createMood
      tags: [moods]
      summary: Log a mood entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoodInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodEntry"
  /subscription:
    get:
      operationId: getSubscription
      tags: [subscription]
      summary: Get the authenticated user's plan and entitlements
      security:
        - bearerAuth: []
      parameters:
        - name: region
          in: query
          required: false
          schema:
            type: string
            enum: [global, indonesia, japan]
      responses:
        "200":
          description: Active plan, localized prices, entitlements, and daily usage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
  /sleep:
    get:
      operationId: getSleepSummary
      tags: [sleep]
      summary: Get the authenticated user's Sleep Support data
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sleep check-ins, journal entries, and routine progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepSummary"
        "503":
          description: Sleep storage is not available
  /sleep/check-ins:
    post:
      operationId: saveSleepCheckIn
      tags: [sleep]
      summary: Save today's sleep check-in
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SleepCheckInInput"
      responses:
        "201":
          description: Saved sleep check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepCheckIn"
        "400":
          description: Invalid check-in
        "503":
          description: Sleep storage is not available
  /sleep/routine-sessions:
    post:
      operationId: saveSleepRoutineSession
      tags: [sleep]
      summary: Save bedtime routine progress
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SleepRoutineSessionInput"
      responses:
        "201":
          description: Saved routine session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepRoutineSession"
        "400":
          description: Invalid routine session
        "503":
          description: Sleep storage is not available
  /sleep/journal:
    post:
      operationId: saveSleepJournalEntry
      tags: [sleep]
      summary: Save a short sleep journal entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SleepJournalEntryInput"
      responses:
        "201":
          description: Saved sleep journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepJournalEntry"
        "400":
          description: Invalid journal entry
        "503":
          description: Sleep storage is not available
  /sleep/insights:
    get:
      operationId: getSleepInsights
      tags: [sleep]
      summary: Get seven-day Sleep Support insights
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sleep insights for Plus and Pro users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepInsights"
        "403":
          description: Plus plan or higher required
        "503":
          description: Sleep storage is not available
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Task:
      type: object
      required: [id, userId, title, status, priority, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
    TaskSummary:
      type: object
      required: [total, pending, in_progress, completed]
      properties:
        total:
          type: integer
        pending:
          type: integer
        in_progress:
          type: integer
        completed:
          type: integer
    MoodEntry:
      type: object
      required: [id, userId, mood, score, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        mood:
          type: string
          enum: [great, good, okay, low, rough]
        score:
          type: integer
        note:
          type: ["string", "null"]
        createdAt:
          type: string
    MoodInput:
      type: object
      required: [mood, score]
      properties:
        mood:
          type: string
          enum: [great, good, okay, low, rough]
        score:
          type: integer
        note:
          type: ["string", "null"]
    SubscriptionStatus:
      type: object
      required: [activePlan, dailyLimit, usedToday, remainingToday, selectedRegion, plans, billingAvailable]
      properties:
        activePlan:
          type: string
          enum: [free, plus, pro]
        dailyLimit:
          type: integer
        usedToday:
          type: integer
        remainingToday:
          type: integer
        selectedRegion:
          type: string
          enum: [global, indonesia, japan]
        billingAvailable:
          type: boolean
        plans:
          type: array
          items:
            $ref: "#/components/schemas/SubscriptionPlan"
    SubscriptionPlan:
      type: object
      required: [id, dailyAiMessages, featureIds, prices]
      properties:
        id:
          type: string
          enum: [free, plus, pro]
        dailyAiMessages:
          type: integer
        featureIds:
          type: array
          items:
            type: string
        prices:
          type: object
          required: [global, indonesia, japan]
          properties:
            global:
              $ref: "#/components/schemas/SubscriptionPrice"
            indonesia:
              $ref: "#/components/schemas/SubscriptionPrice"
            japan:
              $ref: "#/components/schemas/SubscriptionPrice"
    SubscriptionPrice:
      type: object
      required: [currency, amount, display]
      properties:
        currency:
          type: string
        amount:
          type: number
        display:
          type: string
    SleepCheckIn:
      type: object
      required: [id, userId, sleepDate, sleepQuality, energyLevel, feeling, notes, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        sleepDate:
          type: string
          format: date
        sleepQuality:
          type: integer
          minimum: 1
          maximum: 5
        energyLevel:
          type: integer
          minimum: 1
          maximum: 5
        feeling:
          type: string
          enum: [rested, okay, tired, restless]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SleepCheckInInput:
      type: object
      required: [sleepDate, sleepQuality, energyLevel, feeling]
      properties:
        sleepDate:
          type: string
          format: date
        sleepQuality:
          type: integer
          minimum: 1
          maximum: 5
        energyLevel:
          type: integer
          minimum: 1
          maximum: 5
        feeling:
          type: string
          enum: [rested, okay, tired, restless]
        notes:
          type: ["string", "null"]
    SleepJournalEntry:
      type: object
      required: [id, userId, sleepDate, content, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        sleepDate:
          type: string
          format: date
        content:
          type: string
        createdAt:
          type: string
    SleepJournalEntryInput:
      type: object
      required: [sleepDate, content]
      properties:
        sleepDate:
          type: string
          format: date
        content:
          type: string
          minLength: 1
          maxLength: 2000
    SleepRoutineSession:
      type: object
      required: [id, userId, sleepDate, completedSteps, completedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        sleepDate:
          type: string
          format: date
        completedSteps:
          type: array
          items:
            type: string
            enum: [digital_sunset, gratitude, release, breathing, goodnight]
        completedAt:
          type: string
    SleepRoutineSessionInput:
      type: object
      required: [sleepDate, completedSteps]
      properties:
        sleepDate:
          type: string
          format: date
        completedSteps:
          type: array
          items:
            type: string
            enum: [digital_sunset, gratitude, release, breathing, goodnight]
    SleepSummary:
      type: object
      required: [todayCheckIn, todayRoutine, recentCheckIns, recentJournalEntries]
      properties:
        todayCheckIn:
          anyOf:
            - $ref: "#/components/schemas/SleepCheckIn"
            - type: "null"
        todayRoutine:
          anyOf:
            - $ref: "#/components/schemas/SleepRoutineSession"
            - type: "null"
        recentCheckIns:
          type: array
          items:
            $ref: "#/components/schemas/SleepCheckIn"
        recentJournalEntries:
          type: array
          items:
            $ref: "#/components/schemas/SleepJournalEntry"
    SleepInsights:
      type: object
      required: [plan, nightsTracked, averageQuality, averageEnergy, trend]
      properties:
        plan:
          type: string
          enum: [plus, pro]
        nightsTracked:
          type: integer
        averageQuality:
          type: number
        averageEnergy:
          type: number
        trend:
          type: string
          enum: [improving, steady, gentler_start]