# lib/api-zod/src/generated/types/sleepCheckInInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mindful Productivity API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SleepCheckInInputFeeling } from './sleepCheckInInputFeeling';

export interface SleepCheckInInput {
  sleepDate: Date;
  /**
     * @minimum 1
     * @maximum 5
     */
  sleepQuality: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  energyLevel: number;
  feeling: SleepCheckInInputFeeling;
  /** @nullable */
  notes?: string | null;
}