# lib/api-zod/src/generated/types/subscriptionPrice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mindful Productivity API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriptionPrice {
  currency: string;
  amount: number;
  display: string;
}